Split image provider payload helpers
